GPII-1574:  Restructured contents of `src/js` in preparation for creating new components.
